+++ config/configgrpc/configgrpc_test.go
@@ -572,9 +572,10 @@ func TestReceiveOnUnixDomainSocket(t *testing.T) {
 
 func TestContextWithClient(t *testing.T) {
 	testCases := []struct {
-		desc     string
-		input    context.Context
-		expected client.Info
+		desc       string
+		input      context.Context
+		doMetadata bool
+		expected   client.Info
 	}{
 		{
 			desc:     "no peer information, empty client",
@@ -625,10 +626,18 @@ func TestContextWithClient(t *testing.T) {
 			},
 		},
 		{
+			desc: "existing client with metadata, no metadata processing",
+			input: client.NewContext(context.Background(), client.Info{
+				Metadata: map[string][]string{"test-metadata-key": {"test-value"}},
+			}),
+			expected: client.Info{},
+		},
+		{
 			desc: "existing client with metadata",
 			input: client.NewContext(context.Background(), client.Info{
 				Metadata: map[string][]string{"test-metadata-key": {"test-value"}},
 			}),
+			doMetadata: true,
 			expected: client.Info{
 				Metadata: map[string][]string{"test-metadata-key": {"test-value"}},
 			},
@@ -639,6 +648,7 @@ func TestContextWithClient(t *testing.T) {
 				client.NewContext(context.Background(), client.Info{}),
 				metadata.Pairs("test-metadata-key", "test-value"),
 			),
+			doMetadata: true,
 			expected: client.Info{
 				Metadata: map[string][]string{"test-metadata-key": {"test-value"}},
 			},
@@ -646,7 +656,7 @@ func TestContextWithClient(t *testing.T) {
 	}
 	for _, tC := range testCases {
 		t.Run(tC.desc, func(t *testing.T) {
-			cl := client.FromContext(contextWithClient(tC.input))
+			cl := client.FromContext(contextWithClient(tC.input, tC.doMetadata))
 			assert.Equal(t, tC.expected, cl)
 		})
 	}
@@ -669,7 +679,7 @@ func TestStreamInterceptorEnhancesClient(t *testing.T) {
 	}
 
 	// test
-	err := enhanceStreamWithClientInformation(nil, stream, nil, handler)
+	err := enhanceStreamWithClientInformation(false)(nil, stream, nil, handler)
 
 	// verify
 	assert.NoError(t, err)